// karavan/src/main/webapp/src/model/CamelModel.tsx
export class Metadata {
    name: string = '';

    public constructor(init?: Partial<Metadata>) {
        Object.assign(this, init);
    }
}

export class Spec {
    flows: FromStep[] = [];

    public constructor(init?: Partial<Spec>) {
        Object.assign(this, init);
    }
}

export class Integration {
    apiVersion: string = 'camel.apache.org/v1';
    kind: string = 'Integration';
    metadata: Metadata = new Metadata();
    spec: Spec = new Spec();

    public constructor(init?: Partial<Integration>) {
        Object.assign(this, init);
    }

    static createNew(): Integration {
        return new Integration({spec: new Spec({flows: []})})
    }
}

export class CamelElement {
   uuid: string = ''
   dslName: string = ''

   constructor(dslName: string) {
       this.uuid = uuidv4()
       this.dslName = dslName
   }

   hasSteps(): boolean {
      return this.hasOwnProperty('steps');
   }
}
// ...
export class FromStep extends ProcessorStep {
    from: From = new From()

    public constructor(init?: Partial<From>) {
        super('fromStep')
        Object.assign(this, {from: new From({...init})})
    }
}